Issue #ISAICP-2369-2: move helper function to retrieve contextual links to a trait.
